midx: factor out `fill_pack_info()`
When generating a new MIDX, the MIDX machinery must generate a list of
packs which will be contained in the new MIDX. These packs comes from
one of two sources: either from an existing MIDX (if there were already
contained in the MIDX), or from the pack directory (otherwise).

In each case, the MIDX machinery has to fill out a `pack_info` struct
for each pack to be used for bookkeeping purposes.

In anticipation of adding new fields to the `pack_info` struct, factor
out a common routine to zero-initialize and fill out some common fields
for each pack we add.

Signed-off-by: Taylor Blau <me@ttaylorr.com>